app/src/main/java/com/grabber/widget/ui/WidgetActivity.kt
package com.grabber.widget.ui

import android.arch.lifecycle.LifecycleOwner
import android.os.Bundle
import android.support.v7.app.AppCompatActivity
import android.support.v7.widget.Toolbar
import com.grabber.widget.R

class WidgetActivity : AppCompatActivity(), LifecycleOwner {

    override fun onCreate(savedInstanceState: Bundle?) {
        super.onCreate(savedInstanceState)
        setContentView(R.layout.activity_main)

        val toolbar = findViewById<Toolbar>(R.id.toolbar)
        setSupportActionBar(toolbar)

        val intent = intent
        val extras = intent.extras
        title = getString(R.string.widget_settings)

        extras?.let {
            val fragment = WidgetFragment()
            fragment.arguments = extras

            val transaction = supportFragmentManager.beginTransaction()
            transaction.replace(R.id.container, fragment)
            transaction.commit()
        }
    }
}
